moved non UI parts to de.uka.ipd.sdq.errorhandling.core
