+++ bundles/de.uka.ipd.sdq.errorhandling.core/META-INF/MANIFEST.MF
@@ -3,6 +3,8 @@ Bundle-ManifestVersion: 2
 Bundle-Name: SDQ Errorhandling core Plug-in
 Bundle-SymbolicName: de.uka.ipd.sdq.errorhandling.core
 Bundle-Version: 5.2.1.qualifier
+Export-Package: de.uka.ipd.sdq.errorhandling.core
+Require-Bundle: org.eclipse.emf.ecore;bundle-version="2.33.0"
 Bundle-Vendor: sdq.ipd.uka.de
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-17
